Alterações realizadas de idioma, renomeação das classes de manager.
